ttools: graphics coords are now floating point not int
The Surface methods that convert between data and graphics
coordinates now use a Point2D.Double (which has double precision
X and Y members) rather than a Point (which has has int X and Y).
Similarly Axis.graphicsToData.

Although in most cases the coordinates have to be quantised to the
integer pixel grid eventually, that is now left to the
PaperType rather than being done at data->graphics coordinate
conversion time.  This allows the plot layers to make use of
the additional precision if required.

The main important fix that takes advantage of this change at the
current commit is for MultiPointForm, which does plotting of,
possibly scaled, vectors and ellipse markers with variable size and
orientation.  Previously, these suffered from unavoidable rounding
errors in some cases, causing the rendering to go badly wrong when
large scaling factors were applied.  Now, those shapes are plotted
correctly.

+++ ttools/src/main/uk/ac/starlink/ttools/plot2/Axis.java
@@ -49,7 +49,7 @@ public abstract class Axis {
      * @param   g  graphics coordinate
      * @return   data coordinate
      */
-    public abstract double graphicsToData( int g );
+    public abstract double graphicsToData( double g );
 
     /**
      * Returns the data bounds that result from performing an axis zoom
@@ -284,7 +284,7 @@ public abstract class Axis {
             return b_ + a_ * d;
         }
 
-        public double graphicsToData( int g ) {
+        public double graphicsToData( double g ) {
             return ( g - b_ ) * a1_;
         }
 
@@ -338,7 +338,7 @@ public abstract class Axis {
                          : Double.NaN;
         }
 
-        public double graphicsToData( int g ) {
+        public double graphicsToData( double g ) {
             return Math.exp( ( g - b_ ) * a1_ );
         }
